refactor(T04): extract CombatAIAnalysis pure helpers into utils.ts
Addresses Gemini review comments on the T04 test commit:

- Move 7 pure helpers + 2 types out of index.tsx into a sibling
  CombatAIAnalysis/utils.ts; index.tsx now only contains
  buildMatchContext and the React component
- Re-export everything from index.tsx so no external imports break
- Point test imports directly at ../utils (testing a logic module,
  not a React file)
- Remove 4 redundant `as any` casts on makeAdvancedAction calls —
  makeUnit already accepts AnyObj[] for advancedActions
